emg/requirements.txt
pyserial>=3.5
matplotlib>=3.7
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
